mission-control/backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1

# Real-time
python-socketio==5.10.0

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Security (Phase 2)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis
redis==5.0.1

# HTTP Client
httpx==0.26.0

# Utils
python-dotenv==1.0.0
